requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
Pillow==8.0.1
psycopg2
pytz==2020.5
sqlparse==0.4.1
